Improves client interface
It's now possible to call notify with a data structure that represents
a message, rather than having to pass each argument. This encourages
better code style in our consumers, because they can write testable
logic that generates a notification in a separate place to the actual
IO for an invocation.

The `Notification` type was just used internally. Now that it is an
exposed part of the interface, it made sense to tidy up the order of
its arguments.

The existing `notify` function has had its arguments re-ordered to
match the new method. It's now identical to call notify either way,
less the Notification itself.

    notify( <message args> , topicArn [, client] )

    notify( Notification( <message args> ) , topicArn, [, client] )