Improve argument parser for "dev" project
Allows developer to specify the stage to use for configuration,
defaults to dev. This restores the previous (broken) functionality
that was available via environment variables. Generally people will
want to use DEV config, but to test the production configuration it is
now possible to specify PROD with `--config-stage`.

Stage is wired up in the dev project's main, and some redundant
duplication is removed.

Also corrects a couple of little mistakes in the argparser help text.